Update NCC Desk Daily Emails Automation.pyw
Removed functionality to define shift summary range with a txt file so that it could be edited. Now based on the editdate which should be named EntryDate as this marks the date the log entry was created. Any entry created today will be included in the shift summary.
Needs testing to ensure that the query would like a datetime obj vs a string date time var.