Update WO Logbook Tool.py
Major change to function call logic. Added proper closure of program and database when no notifications need to be sent. If notifications are to be sent, then the program will need to be manually closed after the last one is sent.
Logic changed to simplify addition of customers and change of reporting requests.