Hidden email Fix .json -> __init__ + package import
Previous implementation of importing emails via .json file was not working despite a few different approaches. I have moved the variables to a __init__ file which is not shared on git for obvious security reasons.

AHK scripts did not want to WinMaximize my Python windows for whatever reason, so i removed that step. Also one of the windows still doesn't get moved to the correct location. I'd love a AHK expert to tell me how my script is stupid and why its not working.... I thought it was simple.